fix: 기존 오류 수정

+++ monicar-event-hub/src/main/java/org/eventhub/domain/VehicleOnEventCreate.java
@@ -28,7 +28,7 @@ public class VehicleOnEventCreate extends VehicleEventCreateAbs {
 			final long existedSum,
 			final VehicleEventType onEvent,
 			final String rawEventAt,
-			final int sum) {
+			final long sum) {
 		LocalDateTime onTime = validateEventAt(rawEventAt);
 		validateSum(existedSum, sum);
